Extract mask token function from npm datasource to utils
