+++ lib/config/decrypt.js
@@ -1,6 +1,6 @@
 const is = require('@sindresorhus/is');
 const crypto = require('crypto');
-const { maskToken } = require('../datasource/npm');
+const { maskToken } = require('../util/mask');
 
 module.exports = {
   decryptConfig,
