+++ lib/datasource/npm/get.js
@@ -8,9 +8,8 @@ const parse = require('github-url-from-git');
 const { isBase64 } = require('validator');
 
 const hostRules = require('../../util/host-rules');
-
+const { maskToken } = require('../../util/mask');
 const { getNpmrc } = require('./npmrc');
-const { maskToken } = require('./mask');
 
 module.exports = {
   getDependency,
