+++ lib/platform/github/gh-got-wrapper.js
@@ -3,6 +3,7 @@ const ghGot = require('gh-got');
 const delay = require('delay');
 const parseLinkHeader = require('parse-link-header');
 const hostRules = require('../../util/host-rules');
+const { maskToken } = require('../../util/mask');
 
 let cache = {};
 let stats = {};
@@ -194,16 +195,6 @@ async function get(path, options, retries = 5) {
     }
     throw err;
   }
-}
-
-// istanbul ignore next
-function maskToken(token) {
-  if (!token) {
-    return token;
-  }
-  return `${token.substring(0, 2)}${new Array(token.length - 3).join(
-    '*'
-  )}${token.slice(-2)}`;
 }
 
 const helpers = ['get', 'post', 'put', 'patch', 'head', 'delete'];
